Added new examples
