+++ modules/client/examples/03-tokenVoting-client/12-prepare-installation.ts
@@ -1,36 +1,51 @@
 /* MARKDOWN
-### Prepare the installation of a multisig plugin
+---
+title: Prepare Installation
+---
+
+## Prepare the Installation of a Token Voting Plugin
 */
 
 import {
   ContextPlugin,
-  MultisigClient,
   PrepareInstallationStep,
-  MultisigPluginPrepareInstallationParams,
+  TokenVotingClient,
+  TokenVotingPluginPrepareInstallationParams,
+  VotingMode,
 } from "@aragon/sdk-client";
 import { context } from "../00-setup/00-getting-started";
 
 // Instantiate a plugin context from the Aragon OSx SDK context.
 const contextPlugin: ContextPlugin = ContextPlugin.fromContext(context);
-
-// Create an Multisig client.
-const multisigClient: MultisigClient = new MultisigClient(
+// Create an TokenVoting client.
+const tokenVotingClient: TokenVotingClient = new TokenVotingClient(
   contextPlugin,
 );
 
-const installationParams: MultisigPluginPrepareInstallationParams = {
+const installationParams: TokenVotingPluginPrepareInstallationParams = {
   settings: {
     votingSettings: {
-      minApprovals: 5,
-      onlyListed: true
+      supportThreshold: 0.5,
+      minParticipation: 0.5,
+      minDuration: 7200,
+      minProposerVotingPower: BigInt(1),
+      votingMode: VotingMode.STANDARD,
     },
-    members: [
-      "0x1234567890123456789012345678901234567890"
-    ]
+    newToken: {
+      name: "test",
+      decimals: 18,
+      symbol: "TST",
+      balances: [
+        {
+          address: "0x1234567890123456789012345678901234567890",
+          balance: BigInt(10),
+        },
+      ],
+    },
   },
   daoAddressOrEns: "0x1234567890123456789012345678901234567890",
 };
-const steps = multisigClient.methods.prepareInstallation(installationParams);
+const steps = tokenVotingClient.methods.prepareInstallation(installationParams);
 for await (const step of steps) {
   switch (step.key) {
     case PrepareInstallationStep.PREPARING:
